requirements.txt
tensorflow==2.12.0
streamlit==1.45.1
matplotlib==3.7.1
seaborn==0.12.2
pandas==1.5.3
numpy==1.23.5
scikit-learn==1.2.2
Pillow==9.5.0
